brix_array

Module framed

Framed Array

A “framed array” is an array viewed through a frame (or a lens). (framed reads better than lensed or viewed).

The idea here is that we generally want to view an array as a stream of windows – block windows, sliding windows, etc. We do want the possibility of non-unformly sized windows (resampling, groupby, rolling+resampling); so we use the term “frame” rather than windows.

Framed arrays implement a common set of methods (FramedArray) (e.g. all aggregations: reduce/sum/mean/etc.) over different Frames.

Modules

private 🔒

Structs

FramedArray

WholeAxis

A frame over the whole axis. Mirrors Frame::WholeAxis.

Window

A windowed frame (coarsen / block / sliding). Mirrors Frame::Window.

Traits

Framed